fix: use cases optimized

+++ src/blog_app/use_cases/articles.py
@@ -1,4 +1,3 @@
-from typing import Any
 from uuid import UUID
 
 from blog_app.domain.entities.articles import Article, ArticleData
@@ -11,39 +10,41 @@ from blog_app.domain.exceptions.articles import (
 from blog_app.domain.exceptions.users import AuthenticationRequired
 from blog_app.domain.repositories.articles import ArticleRepository
 
+MIN_TITLE_LENGTH = 3
+MIN_CONTENT_LENGTH = 10
+MIN_LOOKING_TEXT_LENGTH = 3
+
+
+def require_authenticated(user: User | None) -> User:
+    if user is None:
+        raise AuthenticationRequired
+    return user
+
+
+def require_admin(user: User) -> None:
+    if not user.is_admin:
+        raise PermissionDenied
+
+
+def validate_article_data(article_data: ArticleData) -> None:
+    if len(article_data.title) < MIN_TITLE_LENGTH:
+        raise TooShortText(f"Title minimal length is {MIN_TITLE_LENGTH} characters")
+
+    if len(article_data.content) < MIN_CONTENT_LENGTH:
+        raise TooShortText(f"Content minimal length is {MIN_CONTENT_LENGTH} characters")
+
+
+def validate_looking_text(text: str) -> None:
+    if len(text) < MIN_LOOKING_TEXT_LENGTH:
+        raise TooShortText(
+            f"Looking text minimal length is {MIN_LOOKING_TEXT_LENGTH} characters"
+        )
+
 
 class BaseArticleUseCase:
     def __init__(self, repo: ArticleRepository, user: User | None) -> None:
         self.repo = repo
-        if user is None:
-            raise AuthenticationRequired
-        self.user = user
-
-
-class ChangeArticleData(BaseArticleUseCase):
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-
-        self.min_title_length = 3
-        self.min_content_length = 10
-
-    def _permission_required(self) -> None:
-        if not self.user.is_admin:
-            raise PermissionDenied
-
-    def _check_fields_length(self, article_data: ArticleData) -> None:
-        title_length = len(article_data.title)
-        content_length = len(article_data.content)
-
-        if title_length < self.min_title_length:
-            raise TooShortText(
-                f"Title minimal length is {self.min_title_length} characters"
-            )
-
-        if content_length < self.min_content_length:
-            raise TooShortText(
-                f"Content minimal length is {self.min_content_length} characters"
-            )
+        self.user = require_authenticated(user)
 
 
 class GetArticleById(BaseArticleUseCase):
@@ -61,29 +62,29 @@ class GetListArticles(BaseArticleUseCase):
         limit_on_page: int | None = None,
         page: int | None = None,
     ) -> list[Article]:
-        if looking_text is not None and len(looking_text) <= 3:
-            raise TooShortText
+        if looking_text is not None:
+            validate_looking_text(looking_text)
         if page and not limit_on_page:
             page = None
         return await self.repo.get_list(looking_text, limit_on_page, page)
 
 
-class CreateArticle(ChangeArticleData):
+class CreateArticle(BaseArticleUseCase):
     async def execute(self, article_data: ArticleData) -> Article:
-        self._permission_required()
-        self._check_fields_length(article_data)
+        require_admin(self.user)
+        validate_article_data(article_data)
         return await self.repo.create(article_data)
 
 
-class UpdateArticle(ChangeArticleData):
+class UpdateArticle(BaseArticleUseCase):
     async def execute(self, article_id: UUID, article_data: ArticleData) -> Article:
-        self._permission_required()
-        self._check_fields_length(article_data)
+        require_admin(self.user)
+        validate_article_data(article_data)
         updated_article = await self.repo.update(article_id, article_data)
         return updated_article
 
 
-class DeleteArticle(ChangeArticleData):
+class DeleteArticle(BaseArticleUseCase):
     async def execute(self, article_id: UUID) -> None:
-        self._permission_required()
+        require_admin(self.user)
         await self.repo.delete(article_id)
